Chargement du dossier client et ajout Socket

+++ server/controller/group.controller.js
@@ -2,6 +2,11 @@ const Group = require("../models/group.model");
 const Membre = require("../models/membre.model");
 const Role = require("../models/role.model");
 
+const { isValidObjectId } = require("mongoose");
+const Conversation = require("../models/conversation.model");
+const Message = require("../models/message.model");
+const { getIO } = require("../socket/socket");
+
 // Middleware pour verifier si l'utilisateur est admin
 const isAdmin = async (userId, groupId) => {
   const membre = await Membre.findOne({
@@ -12,11 +17,36 @@ const isAdmin = async (userId, groupId) => {
   return membre && membre.role_id.type === "admin";
 };
 
+// inclure le nombre de membre dans les groupes
+const countGroupMembres = async (group_id) => {
+  return await Membre.countDocuments({ group_id });
+};
+
 // creation de groupe et ajouter auto l'user createur
 exports.createGroup = async (req, res) => {
   try {
     const { nom, description } = req.body;
-    const createur_id = req.user._id; // USER CONNECTER
+    const createur_id = req.user?._id || req.user?.id; // USER CONNECTER req.user?.id
+
+    if (!req.user || !req.user._id || !req.user.id) {
+      return res.status(401).json({
+        message: "Utilisateur non authentifié",
+      });
+    }
+
+    if (!nom) {
+      return res.status(400).json({
+        error: "Nom du groupe requis",
+      });
+    }
+
+    // verification si le groupe avec un meme nom existe
+    const existingGroup = await Group.findOne({ nom });
+    if (existingGroup) {
+      return res.status(400).json({
+        message: "Un groupe avec ce nom existe déjà",
+      });
+    }
 
     // creation du groupe
     const group = new Group({
@@ -26,7 +56,12 @@ exports.createGroup = async (req, res) => {
       date_creation: new Date(),
     });
 
-    await group.save();
+    const savedGroup = await group.save();
+    if (!savedGroup) {
+      return res.status(404).json({
+        erreur: "Groupe non creer et enregistrer",
+      });
+    }
 
     // attribution de l'user createur comme membre avec role
     const adminRole = await Role.findOne({ type: "admin" });
@@ -38,15 +73,41 @@ exports.createGroup = async (req, res) => {
       date_join: new Date(),
     });
 
+    // creation de conversation de groupe
+    const conversation = await Conversation.create({
+      type: "group",
+      group_id: savedGroup._id,
+    });
+
+    const populatedGroup = await Group.findById(savedGroup._id).populate(
+      "createur_id",
+      "nom pseudo avatar"
+    );
+
+    const io = getIO();
+    // evenement socket pour notifier la creation du groupe
+    if (io) {
+      io.emit("newGroup", { group: savedGroup, createur_id: createur_id });
+    } else {
+      console.error("Socket.IO non initialisé");
+      return res.status(500).json({
+        message:
+          "Erreur serveur interne. Socket.IO non disponible pour la diffusion",
+      });
+    }
+
     res.status(201).json({
       success: true,
       message: "Groupe créé et administrateur ajouté",
-      data: group,
+      data: {
+        group: populatedGroup,
+        conversation: conversation,
+      },
     });
   } catch (error) {
     res.status(500).json({
       succes: false,
-      message: "Erreur lors de la creation du groupe",
+      message: "Erreur du serveur lors de la creation du groupe",
       error: error.message,
     });
   }
@@ -57,10 +118,26 @@ exports.updateGroup = async (req, res) => {
   try {
     const { groupId } = req.params;
     const { nom, description } = req.body;
-    const adminId = req.user._id;
+    const adminId = req.user?._id || req.user?.id; // nosolona le req.user._id req.user?.id
 
+    if (!req.user || !req.user._id || !req.user.id) {
+      return res.status(401).json({
+        message: "Utilisateur non authentifié",
+      });
+    }
+
+    if (!isValidObjectId(adminId)) {
+      return res.status(400).json({
+        message: "ID de l'user admin invalide",
+      });
+    }
+    if (!isValidObjectId(groupId)) {
+      return res.status(400).json({
+        message: "ID du groupe invalide",
+      });
+    }
     //declaration adminRole
-    /*const adminRole = await Role.findOne({ type: "admin" });
+    const adminRole = await Role.findOne({ type: "admin" }).select("_id");
     if (!adminRole) {
       return res.status(400).json({
         success: false,
@@ -71,25 +148,40 @@ exports.updateGroup = async (req, res) => {
     const adminMembre = await Membre.findOne({
       user_id: adminId,
       group_id: groupId,
-      role_id: adminRole._id,
+      role_id: adminRole,
     });
 
     if (!adminMembre) {
       return res.status(403).json({
         success: false,
-        error: "Vous n'êtes pas autorisé à modifier ce groupe",
+        message:
+          "Seuls les administrateurs du groupe peuvent faire cet actions",
       });
-    } */
+    }
 
-    if (!(await isAdmin(adminId, groupId))) {
+    const isUserAdmin = await isAdmin(adminId, groupId);
+    if (!isUserAdmin) {
       return res.status(403).json({
-        error: "Vous n'etes pas autorisé à modifier ce groupe",
+        message: "Vous n'etes pas autorisé à modifier ce groupe",
       });
     }
     // await group.save();
+
+    // Validation des champs
+    if (!nom && !description) {
+      return res.status(400).json({ message: "Aucune modification fournie" });
+    }
+
+    if (!Object.keys(req.body).length) {
+      return res.status(400).json({
+        message: "Aucune donnée fournie pour la mise à jour",
+      });
+    }
+
     const updatedGroup = await Group.findByIdAndUpdate(
       groupId,
-      { nom, description },
+      { ...(nom && { nom }), ...(description && { description }) },
+      //{ nom, description },
       { new: true }
     );
 
@@ -99,15 +191,30 @@ exports.updateGroup = async (req, res) => {
       });
     }
 
+    const io = getIO();
+    // evenement socket.io pour notifier de la mise a jour du groupe
+    if (io) {
+      io.to(`group_${groupId}`).emit("groupModified", {
+        groupId,
+        group: updatedGroup,
+      });
+    } else {
+      console.error("Socket.IO non initialisé");
+      return res.status(500).json({
+        message:
+          "Erreur serveur interne. Socket.IO non disponible pour la diffusion",
+      });
+    }
+
     res.status(200).json({
       success: true,
       message: "Groupe mis à jour avec succès.",
-      group: updatedGroup,
+      data: updatedGroup,
     });
   } catch (error) {
     res.status(500).json({
       success: false,
-      message: "Erreur lors de la modification du groupe",
+      message: "Erreur du serveur lors de la modification du groupe",
       error: error.message,
     });
   }
@@ -117,7 +224,30 @@ exports.updateGroup = async (req, res) => {
 exports.deleteGroup = async (req, res) => {
   try {
     const { groupId } = req.params; // req.params.id
-    const adminId = req.user._id; // req.user.id
+    const adminId = req.user?._id || req.user?.id; // req.user?.id
+
+    if (!req.user || !req.user._id || !req.user.id) {
+      return res.status(401).json({
+        message: "Utilisateur non authentifié",
+      });
+    }
+
+    if (!adminId) {
+      return res.status(404).json({
+        erreur: "Utilisateur non recuperer",
+      });
+    }
+
+    if (!isValidObjectId(groupId)) {
+      return res.status(400).json({
+        message: "ID du group invalide ",
+      });
+    }
+    if (!isValidObjectId(adminId)) {
+      return res.status(400).json({
+        message: "ID user admin invalide ",
+      });
+    }
 
     // Verifie si l'utilisateur est admin du group
     const adminRole = await Role.findOne({ type: "admin" });
@@ -141,18 +271,33 @@ exports.deleteGroup = async (req, res) => {
       });
     }
 
+    if (!(await isAdmin(adminId, groupId))) {
+      return res.status(403).json({
+        message: "Vous n'etes pas autorisé à modifier ce groupe",
+      });
+    }
+
     // supprimer le groupe
     const deletedGroup = await Group.findByIdAndDelete(groupId);
-
     if (!deletedGroup) {
       return res.status(404).json({
         success: false,
         error: "groupe introuvable",
       });
     }
-
     //supprime tous les membres
     await Membre.deleteMany({ group_id: groupId }); // Supprimer tous les membres du groupe
+
+    const io = getIO();
+    if (io) {
+      io.emit("groupRemoved", { groupId, message: "Groupe supprimer" });
+    } else {
+      console.error("Socket.IO non initialisé");
+      return res.status(500).json({
+        message:
+          "Erreur serveur interne. Socket.IO non disponible pour la diffusion",
+      });
+    }
 
     res.status(200).json({
       succes: true,
@@ -161,25 +306,50 @@ exports.deleteGroup = async (req, res) => {
   } catch (error) {
     res.status(500).json({
       success: false,
-      message: "Erreur lors de la suppression du groupe.",
+      message: "Erreur du serveur lors de la suppression du groupe.",
       error: error.message,
     });
   }
 };
 
-// reccuperation de tous les groupes
+// reccuperation de tous les groupes avec comptage dynamique de membres
 exports.getGroups = async (req, res) => {
   try {
-    const groups = await Group.find();
+    const groups = await Group.find().populate("createur_id", "nom");
+
+    if (groups.length === 0 || !groups) {
+      return res.status(404).json({
+        message: "Aucun groupe recuperer",
+      });
+    }
+
+    //Ajout d'un champ membreCount pour chaque groupe
+    // Promise.all: permet de traiter chaque groupe de maniere asynchrone
+    const groupWithMembreCount = await Promise.all(
+      groups.map(async (group) => {
+        const membreCount = await Membre.countDocuments({
+          group_id: group._id,
+        });
+        return {
+          ...group._doc, // inclure tous les champs existant du group
+          TotalMembre: membreCount, // ajouter le champ membreCount
+        };
+      })
+    );
+
+    const count = await Membre.countDocuments(groups);
     res.status(200).json({
       success: true,
-      data: groups,
+      message: "Voici la liste de tous les groupes",
+      //data: groups,
+      groupes: groupWithMembreCount,
+      total: count,
     });
   } catch (error) {
     res.status(500).json({
       success: false,
-      message: "Erreur lors de la récupération des groupes",
-      error,
+      message: "Erreur serveur lors de la récupération des groupes",
+      error: error.message,
     });
   }
 };
@@ -187,9 +357,17 @@ exports.getGroups = async (req, res) => {
 // reccuperation de groupes par ID
 exports.getGroupById = async (req, res) => {
   try {
-    const group = await Group.findById(req.params.id).populate(
+    const { groupId } = req.params;
+
+    if (!isValidObjectId(groupId)) {
+      return res.status(400).json({
+        message: "ID du group invalide ",
+      });
+    }
+
+    const group = await Group.findById(groupId).populate(
       "createur_id",
-      "nom"
+      "nom pseudo avatar"
     );
 
     if (!group) {
@@ -198,228 +376,134 @@ exports.getGroupById = async (req, res) => {
       });
     }
 
+    const membreCount = await countGroupMembres(groupId);
+
     res.status(200).json({
       success: true,
+      message: "Voici le groupe recuperer",
       data: group,
+      membre: membreCount,
     });
   } catch (error) {
     res.status(500).json({
       success: false,
-      message: "Erreur lors de la récupération du groupe",
+      message: "Erreur du serveur lors de la récupération du groupe",
       error: error.message,
     });
   }
 };
 
-//recuperer tout les membres d'un groupe
+//recuperer tout les membres d'un groupe OKEY TSY MILA KITINA
 exports.getGroupMembres = async (req, res) => {
   try {
     const { groupId } = req.params;
-    const membres = await Membre.find({ group_id: groupId })
-      .populate("user_id", "nom")
-      .populate("role_id", "type");
 
-    if (!membres) {
-      return res.status(404).json({
-        error: "Membres introuvable",
+    if (!isValidObjectId(groupId)) {
+      return res.status(400).json({
+        message: "ID du group invalide ",
       });
     }
+    const membres = await Membre.find({ group_id: groupId })
+      .populate("user_id", "nom prenom pseudo avatar")
+      .populate("role_id", "type")
+      .populate("group_id", "nom");
+
+    if (!membres || membres.length === 0) {
+      return res.status(404).json({
+        error: "Aucun membre trouvé pour ce groupe",
+        data: null,
+      });
+    }
+
     res.status(200).json({
       success: true,
+      message: `Liste des membres du groupe récupérée avec succès`,
       data: membres,
-      membres: membres.map((m) => m.user_id),
     });
   } catch (error) {
     res.status(500).json({
       success: false,
-      message: "Erreur lors de la récupération des membres du groupe",
+      message:
+        "Erreur du serveur lors de la récupération des membres du groupe",
       error: error.message,
     });
   }
 };
 
-// recuperer les groupes d'un utilisateur
+// recuperer les groupes auquel l'user connecter est membre
 exports.getUsersGroups = async (req, res) => {
   try {
-    const { userId } = req.params;
-    const groups = await Membre.find({ user_id: userId }).populate("group_id");
+    const userId = req.user || req.user?.id || req.user?._id;
 
-    if (!groups) {
+    if (!req.user || !req.user.id || !req.user._id) {
+      return res.status(401).json({
+        message: "Utilisateur non authentifié ",
+      });
+    }
+
+    if (!isValidObjectId(userId)) {
+      return res.status(400).json({
+        message: "ID de l'user invalide ",
+      });
+    }
+    const groups = await Membre.find({ user_id: userId })
+      .distinct("group_id")
+      .populate("group_id", "nom description")
+      .populate("role_id", "type")
+      .populate("user_id", "nom avatar pseudo");
+
+    if (!groups || groups.length === 0) {
       return res.status(404).json({
         success: false,
-        message: "Groupe(s) introuvable",
+        message: "Utilisateur n'est membre d'aucun groupe",
+      });
+    }
+
+    const groupConversations = await Conversation.find({
+      type: "group",
+      group_id: {
+        $in: groups,
+      },
+    })
+      .populate("group_id", "nom description")
+      .lean();
+
+    for (const conversation of groupConversations) {
+      const lastMessage = await Message.findOne({
+        conversation_id: conversation._id,
+      })
+        .sort({ createdAt: -1 })
+        .select("contenu date_envoi user_id conversation_id")
+        .populate("conversation_id", "type")
+        .populate("user_id", "nom pseudo avatar");
+
+      conversation.dernierMessage = lastMessage || null;
+    }
+
+    if (!groupConversations || groupConversations.length === 0) {
+      return res.status(404).json({
+        message: "Aucune conversation de groupe concernant cet utilisateur",
+      });
+    }
+
+    const io = getIO();
+    if (io) {
+      groupConversations.forEach((group) => {
+        io.to(`user_${userId}`).emit("joinGroup", { groupId: group._id });
       });
     }
 
     res.status(200).json({
       success: true,
-      groups: groups.map((g) => g.group_id),
+      message: `Voici tous les groupes auquel l'utilisateur est membre`,
+      data: groupConversations,
     });
   } catch (error) {
     res.status(500).json({
       success: false,
-      message: "Erreur lors de la récupération des groupes de l'utilisateur",
+      message:
+        "Erreur du serveur lors de la récupération des groupes de l'utilisateur",
       error: error.message,
     });
   }
 };
-
-// recherche de groupe
-exports.searchGroup = async (req, res) => {
-  const { query, page = 1, limit = 10, userId } = req.query; // parametre de recherche
-  const skip = (page - 1) * limit; // nombre de resultat a ignorer pour la page actuelle
-
-  const filters = {};
-  if (userId) filters.createur_id = userId; // Filtre par utilisateur createur de groupe
-
-  try {
-    const groups = await Group.find({
-      $text: { $search: query }, // Recherche full-text sur le nom et la description
-      ...filters, // Applique le filtres
-    })
-      .skip(skip)
-      .limit(parseInt(limit))
-      .exec();
-
-    const totalGroups = await Group.countDocuments({
-      $text: { $search: query },
-      ...filters,
-    });
-
-    res.json({
-      total: totalGroups,
-      page,
-      totalPages: Math.ceil(totalGroups / limit),
-      groups,
-    });
-  } catch (error) {
-    res.status(500).json({
-      succes: false,
-      message: "Erreur serveur lors de la recherche de groupes",
-      error: error.message,
-    });
-  }
-};
-
-/*
-// creer un groupe
-exports.createGroup = async (req, res) => {
-  try {
-    const { nom, description, createur_id } = req.body;
-
-    const newGroup = new Group({
-      nom,
-      description,
-      date_creation: Date.now(),
-      createur_id,
-    });
-    await newGroup.save();
-
-    // Reccuperer le role admin
-    const adminRole = await Role.findOne({ type: "admin" });
-
-    if (!adminRole) {
-      return res.status(400).json({
-        message: "Role admin introuvable",
-      });
-    }
-
-    // ajout du createur comme membre avec role admin
-    const newMember = new Member({
-      user_id: createur_id,
-      group_id: newGroup._id,
-      role_id: adminRole._id,
-    });
-    await newMember.save();
-
-    res.status(201).json(newGroup);
-  } catch (error) {
-    res.status(400).json({
-      message: "Erreur lors de la creation de l'utilisateur",
-      error,
-    });
-  }
-};
-
-
-// Mise a jour du groupe
-exports.updateGroup = async (req, res) => {
-  try {
-    const updateGroup = await Group.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-    });
-
-    if (!updateGroup) {
-      return res.status(404).json({
-        message: "Groupe introuvable",
-      });
-    }
-
-    res.status(200).json(updateGroup);
-  } catch (error) {
-    res.status(400).json({
-      message: "Erreur lors de la mise à jour du groupe",
-      error,
-    });
-  }
-};
-
-// Suppression d'un groupe
-exports.deleteGroup = async (req, res) => {
-  try {
-    const deleteGroup = await group.findByIdAndDelete(req.params.id);
-
-    if (!deleteGroup) {
-      return res.status(404).json({
-        message: "Groupe introuvable",
-      });
-    }
-
-    res.status(200).json({ message: "Groupe supprimé" });
-  } catch (error) {
-    res.status(400).json({
-      message: "Erreur lors de la suppression du groupe",
-      error,
-    });
-  }
-};
-
-
- Ajouter un membre a un groupe
-exports.addMembre = async (req, res) => {
-  const { user_id, group_id, role_id } = req.body;
-  try {
-    const newMembre = newMembre({
-      date_join: Date.now(),
-      user_id,
-      group_id,
-      role_id,
-    });
-
-    await newMembre.save();
-    res.status(201).json(newMembre);
-  } catch (error) {
-    res.status(500).json({
-      message: "Erreur lors de l'ajout du membre",
-      error: error.message,
-    });
-  }
-};
-
- Supprimer un membre d'un groupe
-exports.removeMembre = async (req, res) => {
-  try {
-    const deletedMembre = await Membre.findByIdAndDelete(req.params.membreId);
-    if (!deletedMembre)
-      return res.status(404).json({ message: "Membre non trouvé" });
-    res.status(200).json({
-      message: "Membre supprimé avec succès",
-    });
-  } catch (error) {
-    res.status(500).json({
-      message: "Erreur lors de la suppression du membre",
-      error: error.message,
-    });
-  }
-};
-*/
